feat(mobile): Phase 1 static APK/IPA analysis (pure-Rust, SafeActive, standalone CLI + report bridge)
- Add `mobile` feature (zip+plist optional under it; included in `full`)
- New crates/eggsec/src/mobile/{mod,apk,ipa}.rs: pure-Rust bounded ZIP + AXML/plist manifest/config/static secret scanning for high-signal findings (debuggable, backup, permissions, transport/ATS, exported components, hardcoded secrets, signing, etc.). to_scan_report_data bridge (like wireless).
- CLI: `eggsec mobile <path.{apk,ipa}>` with --json/-o/--quiet; gated command surface + handler.
- Policy: central EnforcementContext via evaluate_and_enforce_operation (SafeActive + required_features:["mobile"]); lab/defense framing only.
- Tests: 15+ new under --features mobile (synthetic ZIPs for APK/IPA cases + ZipSlip + policy handler test).
- Docs: new docs/MOBILE.md; updates to README, CAPABILITIES, SAFETY, FEATURES, AGENTS.md, architecture/{overview,feature_matrix,cli_commands,defense_lab,mobile.md}; mobile AGENTS.override.md stub.
- Incidental latent fixes surfaced by --features full (AtomicUsize import in packet/capture, traceroute u32 max_retries alignment, fuzzer/advanced Severity import, websocket origin mut borrow) — pre-existing, now compile-clean under full.
- All verification (default + mobile + full + clippy + tests) green. Follows existing patterns (wireless/auth-test as standalone defense-lab + report bridge).

See plans/mobile-first-handoff-plan.md (Phase 1 complete).

+++ crates/eggsec/src/packet/cli.rs
@@ -451,7 +451,7 @@ pub async fn handle_packet_traceroute(
         target: args.target.clone(),
         max_hops: args.max_hops,
         timeout: Duration::from_secs(args.timeout.unwrap_or(3)),
-        max_retries: args.probes,
+        max_retries: args.probes as u32,
         first_ttl: args.first_ttl.unwrap_or(1),
         port: 33434,
         use_icmp: use_icmp && is_root(),
